refactor(webbackend): Decouple relations with Persistence and Appliacation layer, refactor Swagger extension
